move new traits, iterators and functions for validation into their own module
